Fix BT CLI unsatisfied imports
