+++ applications/services/bt/bt_service/bt_api.c
@@ -77,39 +77,3 @@ void bt_keys_storage_set_default_path(Bt* bt) {
 
     bt_keys_storage_set_file_path(bt->keys_storage, BT_KEYS_STORAGE_PATH);
 }
-
-/*
- * Private API for the Settings app
- */
-
-void bt_get_settings(Bt* bt, BtSettings* settings) {
-    furi_assert(bt);
-    furi_assert(settings);
-
-    BtMessage message = {
-        .lock = api_lock_alloc_locked(),
-        .type = BtMessageTypeGetSettings,
-        .data.settings = settings,
-    };
-
-    furi_check(
-        furi_message_queue_put(bt->message_queue, &message, FuriWaitForever) == FuriStatusOk);
-
-    api_lock_wait_unlock_and_free(message.lock);
-}
-
-void bt_set_settings(Bt* bt, const BtSettings* settings) {
-    furi_assert(bt);
-    furi_assert(settings);
-
-    BtMessage message = {
-        .lock = api_lock_alloc_locked(),
-        .type = BtMessageTypeSetSettings,
-        .data.csettings = settings,
-    };
-
-    furi_check(
-        furi_message_queue_put(bt->message_queue, &message, FuriWaitForever) == FuriStatusOk);
-
-    api_lock_wait_unlock_and_free(message.lock);
-}
